See commit details.
Since we can’t just monitor the region’s (aka beacon.enteredRegion and beacon.exitedRegion) we need to monitor the beacons.

So once it has been decided that you are on a trip, the code will attempt to determine the closest beacon.

Then we monitor the closest beacon to see if it went out of range.  AKA when didRangeBeacons doesn’t contain the beacon anymore with the same major and minor values.

Once the closest beacon is gone __or__ the region exits we end the trip.